chore: saving

+++ backend/main.py
@@ -3,10 +3,10 @@ from typing import Annotated, List
 from fastapi import FastAPI, HTTPException, Depends, Form, status, File, UploadFile
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import FileResponse
-from models import Image
+from models import Image, User
 from database import db_connect, create_session, get_db
 from sqlalchemy.orm import Session
-from schemas import AdvertisementResponse, AdvertisementCreate
+from schemas import UserBase
 from auth import read_users_me, router
 import os
 import shutil
@@ -19,7 +19,13 @@ session = create_session(engine)
 
 origins = [
     "http://localhost:5173",
-    "http://localhost:3000",
+    "http://localhost:4173",
+    "http://0.0.0.0:4173",
+    "http://0.0.0.0:4173",
+    "http://192.168.100.76:4173/",
+    "http://172.17.0.1:4173/",
+    "http://192.168.100.76:5173",
+    "http://192.168.100.76:4173",
 ]
 
 app.add_middleware(
@@ -39,7 +45,7 @@ async def periodic_cleanup():
     while True:
         with session as db:
             await delete_expired_images(db)
-        await asyncio.sleep(3600)  # Run every hour
+        await asyncio.sleep(30)  # Run every 30 seconds
 
 @app.on_event("startup")
 async def startup_event():
@@ -53,28 +59,24 @@ async def getLogin(user: user_dependency, db: Annotated[Session, Depends(get_db)
     return {"message" : "Welcome Admin!"}
 
 
-
-# @app.get('/get-ads', response_model=List[dict])
-# def get_ad_link(db: Session = Depends(get_db)):
-#     ads = db.query(Ad.media_url, Ad.duration).all()
-#     list_of_ad = []
-#     for ad in ads:
-#         list_of_ad.append({"url":get_embed_url(ad.media_url), "duration":ad.duration})
-#     print(list_of_ad)
-#     return list_of_ad
-
-
-
+# Upload Images and store the filename and filepath in the database
 @app.post("/upload/")
-async def upload_file(file: UploadFile = File(...), db=Depends(get_db)):
+async def upload_file(file: UploadFile = File(...), current_user: UserBase = Depends(read_users_me) ,db: Session = Depends(get_db)):
+    if not file:
+        raise HTTPException(status_code=400, detail="No file uploaded")
+    
+    file_exist = db.query(Image.filename).filter(Image.filename == file.filename).first()
+    if file_exist:
+        raise HTTPException(status_code=400, detail="File already exists")
+    
     file_path = os.path.join(UPLOAD_DIR, file.filename)
     
     with open(file_path, "wb") as buffer:
         shutil.copyfileobj(file.file, buffer)
     
-    expiration_time = datetime.now() + timedelta(days=1)  # 1-day expiration
+    expiration_time = datetime.now() + timedelta(minutes=1)  # 1-day expiration
 
-    new_image = Image(filepath=file.filename, duration=15, expires_in=expiration_time)
+    new_image = Image(filename=file.filename, duration=15, expires_in=expiration_time)
     db.add(new_image)
     db.commit()
     db.refresh(new_image)
@@ -86,24 +88,30 @@ async def delete_expired_images(db: Session):   # Delete expired images
     expired_images = db.query(Image).filter(Image.expires_in < now).all()
     
     for image in expired_images:
-        if os.path.exists(image.filepath):
-            os.remove(image.filepath)
-        db.delete(image)
+        if os.path.exists(os.path.join(UPLOAD_DIR, image.filename)):
+            try:
+                db.delete(image)
+            except OSError as e:
+                print(f"Error deleting file: {e}")
+            finally:
+                os.remove(os.path.join(UPLOAD_DIR, image.filename))
+        
     
     db.commit()
     
     return {"message": f"Deleted {len(expired_images)} expired images."}
 
 
-# TODO: Get the images on the database
+# Get all images filename
 @app.get('/')
 async def get_images(db: Session = Depends(get_db)):
-    images = db.query(Image.filepath).all()
+    images = db.query(Image.filename).all()
     list_of_images = []
     for image in images:
-        list_of_images.append(image.filepath)
+        list_of_images.append(image.filename)
     return list_of_images
 
+# GET images
 @app.get("/media/{filename}")
 async def get_file(filename: str):
     file_path = os.path.join(UPLOAD_DIR, filename)
